+++ app/Http/Controllers/Back/LinkController.php
@@ -5,18 +5,14 @@ namespace App\Http\Controllers\Back;
 use App\Http\Controllers\Controller;
 use App\Models\Link;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Storage;
 
 class LinkController extends Controller
 {
-    public function __construct()
-    {
-        $this->authorizeResource(Link::class, 'link');
-    }
-
     public function index()
     {
-        $groups = config('front.linkGroups', []);
+        $this->authorize('links.index');
+
+        $groups = config('general.linkGroups', []);
         $links = Link::orderBy('ordering')->get();
 
         return view('back.links.index', compact('groups', 'links'));
@@ -24,24 +20,28 @@ class LinkController extends Controller
 
     public function create()
     {
-        $groups = config('front.linkGroups', []);
+        $this->authorize('links.create');
+
+        $groups = config('general.linkGroups', []);
 
         return view('back.links.create', compact('groups'));
     }
 
     public function store(Request $request)
     {
+        $this->authorize('links.create');
+
         $this->validate($request, [
-            'title'         => 'required',
-            'link'          => 'required',
+            'title' => 'required',
+            'url' => 'required',
             'link_group_id' => 'required',
         ]);
 
         Link::create([
-            'title'            => $request->title,
-            'link'             => $request->link,
-            'link_group_id'    => $request->link_group_id,
-            'ordering'         => Link::max('ordering') + 1,
+            'title' => $request->title,
+            'url' => $request->url,
+            'link_group_id' => $request->link_group_id,
+            'ordering' => Link::max('ordering') + 1,
         ]);
 
         toastr()->success('لینک با موفقیت ایجاد شد.');
@@ -51,23 +51,27 @@ class LinkController extends Controller
 
     public function edit(Link $link)
     {
-        $groups = config('front.linkGroups', []);
+        $this->authorize('links.update');
+
+        $groups = config('general.linkGroups', []);
 
         return view('back.links.edit', compact('link', 'groups'));
     }
 
     public function update(Link $link, Request $request)
     {
+        $this->authorize('links.update');
+
         $this->validate($request, [
-            'title'         => 'required',
-            'link'          => 'required',
+            'title' => 'required',
+            'url' => 'required',
             'link_group_id' => 'required',
         ]);
 
         $link->update([
-            'title'            => $request->title,
-            'link'             => $request->link,
-            'link_group_id'    => $request->link_group_id,
+            'title' => $request->title,
+            'url' => $request->url,
+            'link_group_id' => $request->link_group_id,
         ]);
 
         toastr()->success('لینک با موفقیت ویرایش شد.');
@@ -77,9 +81,7 @@ class LinkController extends Controller
 
     public function destroy(Link $link)
     {
-        if ($link->image) {
-            Storage::disk('public')->delete($link->image);
-        }
+        $this->authorize('links.delete');
 
         $link->delete();
 
@@ -95,7 +97,6 @@ class LinkController extends Controller
         ]);
 
         $i = 1;
-
         foreach ($request->links as $link) {
             Link::findOrFail($link)->update([
                 'ordering' => $i++,
@@ -105,20 +106,18 @@ class LinkController extends Controller
         return response('success');
     }
 
-    // groups methods
-
     public function groups()
     {
         $this->authorize('links.groups');
 
-        $groups = config('front.linkGroups', []);
+        $groups = config('general.linkGroups', []);
 
         return view('back.links.groups', compact('groups'));
     }
 
     public function updateGroups(Request $request)
     {
-        $this->authorize('links.groups');
+        $this->authorize('links.groups.update');
 
         $request->validate([
             'groups' => 'required|array',
